requirements.txt
crewai-tools[mcp]
mcp
uv
python-dotenv
langchain-mcp-adapters
langchain [openai]
pymongo
langchain-mongodb
lida
llmx openai